requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
lightgbm==4.3.0
shap==0.46.0
matplotlib==3.9.0
pyyaml==6.0.2
